refactor(server): split mobility cleanup service
