+++ scripts/max_lines_allowlist.txt
@@ -34,7 +34,7 @@ server/proliferate/db/store/cloud_sync/commands.py 1565 server structure migrati
 server/proliferate/db/store/cloud_sync/events.py 1109 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_workspaces.py 1187 server structure migration debt for oversized store file
 server/proliferate/server/billing/service.py 2459 server structure migration debt for oversized service file during billing orchestration consolidation
-server/proliferate/server/cloud/mobility/service.py 1208 server structure migration debt for oversized service file
+server/proliferate/server/cloud/mobility/service.py 904 server structure migration debt for oversized service file
 server/tests/integration/schema_migration_assertions.py 878 existing server oversized test helper
 server/tests/integration/test_auth_flow.py 2057 existing server oversized integration test
 server/tests/integration/test_billing_accounting.py 871 existing server oversized integration test
