+++ server/proliferate/server/cloud/mobility/cleanup/__init__.py
@@ -1,0 +1,1 @@
+"""Workspace mobility cleanup subdomain."""
