+++ server/proliferate/server/cloud/mobility/cleanup/domain/__init__.py
@@ -1,0 +1,1 @@
+"""Pure workspace mobility cleanup rules."""
